Move class definition of Gas voronoi table
